LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/form-double-submit-preventdefault-expected.txt





FAIL PreventDefaulting form onsubmit should allow submit() to succeed assert_equals: expected "v6" but got "v7"
FAIL PreventDefaulting form onsubmit should allow submit() to succeed and the second submit() which is invalid should not supersede first one assert_equals: expected "v2" but got "v4"
